solution/requirements.txt
pyodbc
psycopg2
Werkzeug>=2.0
flask-restx>=0.5.0
sqlalchemy==1.3.20
Flask-SQLAlchemy==2.4.4
pandas
requests
pydantic
openpyxl